Move list page to /forms/user-applications
